Moving out recursive minimisation
