+++ pumpkin-lib/src/engine/conflict_analysis/mod.rs
@@ -1,9 +1,11 @@
 //! Contains algorithms for conflict analysis, core extraction, and clause minimisation.
 //! The algorithms use resolution and implement the 1uip and all decision literal learning schemes
 mod conflict_analysis_context;
+mod recursive_minimisation;
 mod resolution_conflict_analyser;
 mod semantic_minimiser;
 
 pub(crate) use conflict_analysis_context::ConflictAnalysisContext;
+pub(crate) use recursive_minimisation::*;
 pub(crate) use resolution_conflict_analyser::*;
 pub(crate) use semantic_minimiser::*;
